factchecker/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
transformers
langchain
numpy
faiss-cpu
requests
easyocr
langchain-community
python-dotenv
tiktoken
python-multipart
openai
torch
tensorflow
flax